+++ .gitignore
@@ -1,0 +1,4 @@
+controller/
+views/
+__pycache__/
+output/
